+++ metta/rl/functions/batch_utils.py
@@ -1,0 +1,37 @@
+"""Batch size and sampling utilities for Metta training."""
+
+from typing import Tuple
+
+
+def calculate_batch_sizes(
+    forward_pass_minibatch_target_size: int,
+    num_agents: int,
+    num_workers: int,
+    async_factor: int,
+) -> Tuple[int, int, int]:
+    """Calculate target batch size, actual batch size, and number of environments.
+
+    Returns:
+        Tuple of (target_batch_size, batch_size, num_envs)
+    """
+    target_batch_size = forward_pass_minibatch_target_size // num_agents
+    if target_batch_size < max(2, num_workers):  # pufferlib bug requires batch size >= 2
+        target_batch_size = num_workers
+
+    batch_size = (target_batch_size // num_workers) * num_workers
+    num_envs = batch_size * async_factor
+
+    return target_batch_size, batch_size, num_envs
+
+
+def calculate_prioritized_sampling_params(
+    epoch: int,
+    total_timesteps: int,
+    batch_size: int,
+    prio_alpha: float,
+    prio_beta0: float,
+) -> float:
+    """Calculate annealed beta for prioritized experience replay."""
+    total_epochs = max(1, total_timesteps // batch_size)
+    anneal_beta = prio_beta0 + (1 - prio_beta0) * prio_alpha * epoch / total_epochs
+    return anneal_beta
