+++ metta/rl/functions/utils.py
@@ -1,0 +1,17 @@
+"""General utilities for Metta training."""
+
+
+def should_run(
+    epoch: int,
+    interval: int,
+    is_master: bool = True,
+    force: bool = False,
+) -> bool:
+    """Check if a periodic task should run based on interval and master status."""
+    if not is_master or not interval:
+        return False
+
+    if force:
+        return True
+
+    return epoch % interval == 0
